requirements.txt
flask==2.0.1
flask-restx==0.5.1
Flask-Migrate==3.1.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
pylint_flask_sqlalchemy
pylint-flask==0.6
Faker==8.14.1
gunicorn==20.1.0
prometheus-flask-exporter==0.18.3
uuid==1.30
mmh3==3.0.0